+++ src/main/java/co/edu/unicauca/asae/cleanarquitecture/infraestructura/output/persistencia/entidades/FormatoAEntity.java
@@ -26,7 +26,7 @@ import lombok.Data;
 public class FormatoAEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idFormatoA;
+    private Long idFormatoA;
 
     @Column(length = 100, unique = true)
     private String titulo;
